Remove Fetch from [id].js
