+++ app/lib/navbar.js
@@ -1,0 +1,8 @@
+import { fetchAPI } from "./api";
+
+export const getNavItems = async ()=>{
+
+    const topNavItems = await fetchAPI("/top-nav-item");
+    return topNavItems;
+
+}
